Extracted calendar module and optimized end of month UX (#483)
